+++ RingGrids/src/RingGrids.jl
@@ -126,6 +126,8 @@ export interpolate,
 # STATISTICS
 export zonal_mean
 
+export unmasked_indices, copy_unmasked!
+
 # CONSTANTS
 const DEFAULT_NF = Float32
 const DEFAULT_ARRAYTYPE = Array
@@ -168,7 +170,7 @@ include("quadrature_weights.jl")
 include("interpolation.jl")
 include("vertices.jl")
 include("statistics.jl")
-include("masked_copy.jl")
+include("copy_unmasked.jl")
 
 # ASSET DOWNLOADING
 export get_asset, load_asset, ASSETS_URL, DEFAULT_ASSETS_VERSION